Add files via upload
Together with Filippo Samorì and Matteo Bonucci, we developed this project as part of the final examination for the Computer Vision and Image Processing course in the Automation Engineering program at the University of Bologna.

Our project focuses on designing an algorithm to detect barcodes in gray-level images from a specific dataset. The algorithm identifies the Region of Interest (ROI) and computes various quality parameters, which are then recorded in an Excel table.

For further details, please refer to the attached PDF report.